refactor: 장바구니 관련 로직 cartUtils 파일로 분리
